restructure common includes
* rename AlpakaThreadElemsDistribution to alpakaHelpers
* inline demangle.hpp into simpletest
* remove Dummy
