--- examples/common/Dummy.cpp
@@ -1,6 +1,0 @@
-#include <boost/core/ignore_unused.hpp>
-
-void dummy(void * d)
-{
-    boost::ignore_unused(d);
-}
