--- examples/common/Dummy.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-void dummy(void * d);
