+++ examples/common/alpakaHelpers.hpp
