+++ examples/simpletest/simpletest.cpp
@@ -11,10 +11,9 @@
  *  domain, to create a view and to access the data
  */
 
-#include "../common/demangle.hpp"
-
 #include <iostream>
 #include <llama/llama.hpp>
+#include <sstream>
 #include <utility>
 #include <vector>
 
@@ -58,13 +57,68 @@ using Name = llama::DS<
     llama::DE<st::Options, llama::DA<bool, 4>>>;
 // clang-format on
 
-/** Prints the coordinates of a given \ref llama::DatumCoord for debugging and
- *  testing purposes
- */
-template<std::size_t... T_coords>
-void printCoords(llama::DatumCoord<T_coords...> dc)
+namespace
 {
-    (std::cout << ... << T_coords);
+    template<class T>
+    std::string type(const T & t)
+    {
+        return boost::core::demangle(typeid(t).name());
+    }
+
+    /** Prints the coordinates of a given \ref llama::DatumCoord for debugging
+     * and testing purposes
+     */
+    template<std::size_t... T_coords>
+    void printCoords(llama::DatumCoord<T_coords...> dc)
+    {
+        (std::cout << ... << T_coords);
+    }
+
+    template<typename Out>
+    void split(const std::string & s, char delim, Out result)
+    {
+        std::stringstream ss(s);
+        std::string item;
+        while(std::getline(ss, item, delim))
+        {
+            *(result++) = item;
+        }
+    }
+
+    std::vector<std::string> split(const std::string & s, char delim)
+    {
+        std::vector<std::string> elems;
+        split(s, delim, std::back_inserter(elems));
+        return elems;
+    }
+
+    std::string nSpaces(int n)
+    {
+        std::string result = "";
+        for(int i = 0; i < n; ++i) result += " ";
+        return result;
+    }
+
+    std::string addLineBreaks(std::string raw)
+    {
+        boost::replace_all(raw, "<", "<\n");
+        boost::replace_all(raw, ", ", ",\n");
+        boost::replace_all(raw, " >", ">");
+        boost::replace_all(raw, ">", "\n>");
+        auto tokens = split(raw, '\n');
+        std::string result = "";
+        int indent = 0;
+        for(auto t : tokens)
+        {
+            if(t.back() == '>' || (t.length() > 1 && t[t.length() - 2] == '>'))
+                indent -= 4;
+            result += nSpaces(indent) + t + "\n";
+            if(t.back() == '<')
+                indent += 4;
+        }
+        return result;
+    }
+
 }
 
 /** Example functor for \ref llama::ForEach which can also be used to print the
